% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{beverage_revenues}
\alias{beverage_revenues}
\title{Beverage revenues}
\format{
A tibble with 4 rows and 3 variables:
\describe{
\item{year}{Calendar year.}
\item{spritzi}{Revenue of Spritzi in thousand EUR.}
\item{prickli}{Revenue of Prickli in thousand EUR.}
}
}
\usage{
beverage_revenues
}
\description{
Annual revenues of two beverage manufacturers in thousand EUR.
}
\keyword{datasets}
