% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_sites_muenster.R
\docType{data}
\name{bike_sites_muenster}
\alias{bike_sites_muenster}
\title{Bike counting sites in Muenster}
\format{
A tibble with 24 rows and 2 variables:
\describe{
\item{station_id}{Station identifier used in the source repository.}
\item{station_name}{Readable station name.}
}
}
\usage{
bike_sites_muenster
}
\description{
Metadata for selected public bike counting stations in Muenster.
}
\keyword{datasets}
