% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvb_rankings.R
\docType{data}
\name{bvb_rankings}
\alias{bvb_rankings}
\title{Borussia Dortmund final rankings (1988-2022)}
\format{
A tibble with 35 rows and 2 variables:
\describe{
\item{year}{Calendar year of the season endpoint.}
\item{ranking}{Final Bundesliga ranking (1 = best rank).}
}
}
\usage{
bvb_rankings
}
\description{
The dataset contains Borussia Dortmund's final Bundesliga ranking for each
season from 1988 to 2022.
}
\keyword{datasets}
