% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_brand_by_occupation_individual.R
\docType{data}
\name{car_brand_by_occupation_individual}
\alias{car_brand_by_occupation_individual}
\title{Car brand by occupation (individual-level)}
\format{
A tibble with 500 rows and 2 variables:
\describe{
\item{occupation}{Occupation category.}
\item{brand}{Car brand category.}
}
}
\usage{
car_brand_by_occupation_individual
}
\description{
The dataset contains one row per surveyed person with occupation and car
brand.
}
\keyword{datasets}
