% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_occupation.R
\docType{data}
\name{car_occupation}
\alias{car_occupation}
\title{Car brand by occupation (aggregated)}
\format{
A tibble with 4 rows and 6 variables:
\describe{
\item{occupation}{Occupation category.}
\item{audi}{Count of Audi drivers.}
\item{bmw}{Count of BMW drivers.}
\item{opel}{Count of Opel drivers.}
\item{vw}{Count of VW drivers.}
\item{total}{Row total.}
}
}
\usage{
car_occupation
}
\description{
The dataset contains counts of preferred car brands by occupation.
}
\keyword{datasets}
