% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout_service_times.R
\docType{data}
\name{checkout_service_times}
\alias{checkout_service_times}
\title{Checkout service times}
\format{
A tibble with 50 rows and 1 variable:
\describe{
\item{service_time_seconds}{Observed service time in seconds.}
}
}
\usage{
checkout_service_times
}
\description{
The dataset contains service times (in seconds) of 50 consecutive customers
at a supermarket checkout.
}
\keyword{datasets}
