% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{cinema_group_size_pmf}
\alias{cinema_group_size_pmf}
\title{Cinema group size probability function}
\format{
A tibble with 5 rows and 2 variables:
\describe{
\item{group_size}{Number of persons per group.}
\item{probability}{Probability mass.}
}
}
\usage{
cinema_group_size_pmf
}
\description{
Probability function for group size in cinema visits.
}
\keyword{datasets}
