% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinema_visitors_100days.R
\docType{data}
\name{cinema_visitors_100days}
\alias{cinema_visitors_100days}
\title{Cinema visitors over 100 days}
\format{
A tibble with 11 rows and 2 variables:
\describe{
\item{viewers}{Number of visitors counted on a day.}
\item{days}{Number of days with the respective visitor count.}
}
}
\usage{
cinema_visitors_100days
}
\description{
The dataset contains an absolute frequency distribution of cinema visitors per
day recorded over 100 days.
}
\keyword{datasets}
