% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{cinema_visitors_ind}
\alias{cinema_visitors_ind}
\title{Cinema visitors over 100 days (individual-level)}
\format{
A tibble with 100 rows and 1 variable:
\describe{
\item{viewers}{Number of visitors counted on the day.}
}
}
\usage{
cinema_visitors_ind
}
\description{
Expanded version of \link{cinema_visitors} with one row per observed day.
}
\keyword{datasets}
