% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{cold_rents}
\alias{cold_rents}
\title{Monthly cold rents}
\format{
A tibble with 10 rows and 1 variable:
\describe{
\item{monthly_rent_eur}{Monthly cold rent in EUR.}
}
}
\usage{
cold_rents
}
\description{
Ten observed monthly cold rents in EUR from one neighborhood.
}
\keyword{datasets}
