% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{company_financials}
\alias{company_financials}
\title{Company profile sample}
\format{
A tibble with 4 rows and 5 variables:
\describe{
\item{company}{Company name.}
\item{employees}{Number of employees.}
\item{annual_revenue_mio_eur}{Annual revenue in million EUR.}
\item{equity_share_pct}{Equity share in percent.}
\item{credit_rating}{Rating category.}
}
}
\usage{
company_financials
}
\description{
Sample of four companies used in introductory tasks on variables and scales.
}
\keyword{datasets}
