% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{computer_sales_country}
\alias{computer_sales_country}
\title{Monthly computer sales in countries A and B}
\format{
A tibble with 7 rows and 3 variables:
\describe{
\item{units}{Monthly number of sold units.}
\item{probability_country_a}{Probability in country A.}
\item{probability_country_b}{Probability in country B.}
}
}
\usage{
computer_sales_country
}
\description{
Two discrete probability distributions for monthly sales volumes.
}
\keyword{datasets}
