% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{credit_ratings_two_reviewers}
\alias{credit_ratings_two_reviewers}
\title{Credit ratings from two reviewers}
\format{
A tibble with 7 rows and 3 variables:
\describe{
\item{company_id}{Company identifier.}
\item{reviewer_a}{Score from reviewer A.}
\item{reviewer_b}{Score from reviewer B.}
}
}
\usage{
credit_ratings_two_reviewers
}
\description{
Two independent reviewers rated seven companies on a 1-10 scale.
}
\keyword{datasets}
