% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_elections.R
\docType{data}
\name{de_elections}
\alias{de_elections}
\title{Bundestag election results by federal state}
\format{
A tibble with 377 rows and 6 variables:
\describe{
\item{election_year}{Election year.}
\item{state_code}{Two-digit federal state code.}
\item{state}{Federal state name.}
\item{party}{Party label in the official result file.}
\item{votes}{Number of second votes.}
\item{vote_share}{Second-vote share in percent.}
}
}
\usage{
de_elections
}
\description{
Party-level second-vote results by German federal state for the 2021
Bundestag election result table (official publication format).
}
\keyword{datasets}
