% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{decathlon_3disc_summary}
\alias{decathlon_3disc_summary}
\title{Decathlon summary for three disciplines}
\format{
A tibble with 3 rows and 3 variables:
\describe{
\item{discipline}{Discipline name.}
\item{mean}{Arithmetic mean of results.}
\item{variance}{Variance of results.}
}
}
\usage{
decathlon_3disc_summary
}
\description{
Mean and variance for 100m, long jump, and shot put from a large decathlon sample.
}
\keyword{datasets}
