% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_survey_eu10.R
\docType{data}
\name{environment_survey_eu10}
\alias{environment_survey_eu10}
\title{Environmental survey sample (EU, n = 10)}
\format{
A tibble with 10 rows and 4 variables:
\describe{
\item{age}{Age of respondent.}
\item{gender}{Gender of respondent.}
\item{income_eur}{Monthly net income in EUR.}
\item{environmental_protection}{Attitude towards environmental protection.}
}
}
\usage{
environment_survey_eu10
}
\description{
The dataset contains responses of 10 sampled persons from an EU-wide survey
on environmental protection.
}
\keyword{datasets}
