% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{exam_study_time}
\alias{exam_study_time}
\title{Exam points and study time}
\format{
A tibble with 5 rows and 3 variables:
\describe{
\item{study_days}{Study time in days.}
\item{points}{Achieved points.}
\item{max_points}{Maximum achievable points.}
}
}
\usage{
exam_study_time
}
\description{
Study days and exam points of five participants.
}
\keyword{datasets}
