% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_sentencing_ind.R
\docType{data}
\name{florida_sentencing_ind}
\alias{florida_sentencing_ind}
\title{Florida murder sentencing by defendant skin color (individual-level)}
\format{
A tibble with 326 rows and 2 variables:
\describe{
\item{defendant_skin_color}{Defendant skin color.}
\item{sentence}{Sentence category.}
}
}
\usage{
florida_sentencing_ind
}
\description{
The dataset contains one row per case with defendant skin color and sentence
category.
}
\keyword{datasets}
