% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{gas_prices}
\alias{gas_prices}
\title{Motorway gas prices}
\format{
A tibble with 10 rows and 1 variable:
\describe{
\item{price_eur_per_liter}{Observed price in EUR per liter.}
}
}
\usage{
gas_prices
}
\description{
Ten observed prices for regular gasoline (EUR per liter) from a motorway
section.
}
\keyword{datasets}
