% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_data.R
\docType{data}
\name{hr_data}
\alias{hr_data}
\title{HR data}
\format{
A tibble with 1,200 rows and 6 variables:
\describe{
\item{position [fct]}{Position of the employee in the company.}
\item{working_hours [dbl]}{Contracted working hours of the employee per week.}
\item{salary [dbl]}{Monthly salary of the employee. }
\item{hourly_wage [dbl]}{Hourly compensation of the employee. }
\item{department [fct]}{Department in which the employee is employed.}
\item{sick_days [dbl]}{Sick days of the employee in the period observed.}
}
}
\usage{
hr_data
}
\description{
The dataset contains human resources data of a footwear company. Each entity of the dataset represents one employee with a total of six attributes.
}
\keyword{datasets}
