% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{library_overdue_model}
\alias{library_overdue_model}
\title{Library overdue model}
\format{
A tibble with 5 rows and 2 variables:
\describe{
\item{weeks_overdue}{Overdue weeks.}
\item{probability_expression}{Probability expression as text.}
}
}
\usage{
library_overdue_model
}
\description{
Symbolic probability model for overdue weeks.
}
\keyword{datasets}
