% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_expenses.R
\docType{data}
\name{marketing_expenses}
\alias{marketing_expenses}
\title{Data of Marketing expenses}
\format{
A tibble with 235 rows and 14 variables:
\describe{
\item{dbl marketing_expenses}{...}
\item{dbl customers_reached}{...}
\item{dbl negative_reactions}{...}
\item{dbl price}{...}
\item{fct price_segment}{...}
\item{dbl number_of_sizes}{...}
\item{fctl target_customer}{...}
\item{dbl rating_testers}{...}
\item{dbl rating_customers}{...}
\item{fct color_most_sold}{...}
\item{dbl return_rate}{...}
\item{dbl sales_volume}{...}
\item{dbl rank_rating_customers}{...}
\item{dbl rank_price_segment}{...}
}
}
\usage{
marketing_expenses
}
\description{
Marketing expenses for shoes.
}
\keyword{datasets}
