% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_bike_sites.R
\docType{data}
\name{ms_bike_sites}
\alias{ms_bike_sites}
\title{Bike counting sites in Muenster}
\format{
A tibble with 24 rows and 2 variables:
\describe{
\item{station_id}{Station identifier used in the source repository.}
\item{station_name}{Readable station name.}
}
}
\usage{
ms_bike_sites
}
\description{
Metadata for selected public bike counting stations in Muenster.
}
\keyword{datasets}
