% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSBStatsData-package.R
\docType{package}
\name{MSBStatsData-package}
\alias{MSBStatsData}
\alias{MSBStatsData-package}
\title{MSBStatsData: Package containing data for my statistics courses at the Münster School of Business}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Bücker \email{michael.buecker@fh-muenster.de} (\href{https://orcid.org/0000-0003-0045-8460}{ORCID})

}
\keyword{internal}
