% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrw_population.R
\docType{data}
\name{nrw_population}
\alias{nrw_population}
\title{Population by NRW districts}
\format{
A tibble with 318 rows and 5 variables:
\describe{
\item{year}{Reference year.}
\item{nuts3_code}{NUTS-3 district code.}
\item{district_name}{District name.}
\item{district_type}{District type (urban or rural district).}
\item{population}{Total resident population.}
}
}
\usage{
nrw_population
}
\description{
Annual population counts for NUTS-3 districts in North Rhine-Westphalia
(NRW), Germany.
}
\keyword{datasets}
