% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{order_quantities}
\alias{order_quantities}
\title{Order quantities}
\format{
A tibble with 20 rows and 1 variable:
\describe{
\item{order_quantity}{Ordered units per order.}
}
}
\usage{
order_quantities
}
\description{
Order quantities per transaction from a small ordering sample.
}
\keyword{datasets}
