% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{populations_i_ii}
\alias{populations_i_ii}
\title{Two populations I and II}
\format{
A tibble with 4 rows and 2 variables:
\describe{
\item{population_i}{Values of population I.}
\item{population_ii}{Values of population II.}
}
}
\usage{
populations_i_ii
}
\description{
Two small populations used for comparing dispersion measures.
}
\keyword{datasets}
