% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{property_prices_distance}
\alias{property_prices_distance}
\title{Property prices and station distance}
\format{
A tibble with 5 rows and 2 variables:
\describe{
\item{distance_km}{Distance to station in km.}
\item{price_eur_per_m2}{Price in EUR per square meter.}
}
}
\usage{
property_prices_distance
}
\description{
Property prices per square meter and distance to next suburban rail station.
}
\keyword{datasets}
