% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{rent_living_area}
\alias{rent_living_area}
\title{Rent and living area}
\format{
A tibble with 5 rows and 2 variables:
\describe{
\item{living_area_m2}{Living area in square meters.}
\item{cold_rent_eur}{Monthly cold rent in EUR.}
}
}
\usage{
rent_living_area
}
\description{
Monthly cold rent and living area for five apartments.
}
\keyword{datasets}
