% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_ad_spending.R
\docType{data}
\name{research_ad_spending}
\alias{research_ad_spending}
\title{Monthly research and advertising spending}
\format{
A tibble with 12 rows and 3 variables:
\describe{
\item{month}{Month abbreviation.}
\item{research}{Research spending in thousand EUR.}
\item{advertising}{Advertising spending in thousand EUR.}
}
}
\usage{
research_ad_spending
}
\description{
The dataset contains monthly spending of a company on research and
advertising (in thousand EUR).
}
\keyword{datasets}
