% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_trends.R
\docType{data}
\name{salary_trends}
\alias{salary_trends}
\title{Salary trend data}
\format{
A tibble with 10 rows and 2 variables:
\describe{
\item{year [dbl]}{Year of record.}
\item{avg_salary [dbl]}{Average salary in the corresponding year.}
}
}
\usage{
salary_trends
}
\description{
Dataset containing the average salary of a footwear company's employees over 10 years.
}
\keyword{datasets}
