% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipping_errors.R
\docType{data}
\name{shipping_errors}
\alias{shipping_errors}
\title{Shipping error data}
\format{
A tibble with 10 rows and 2 variables:
\describe{
\item{error [fct]}{...}
\item{shipping [fct]}{...}
}
}
\usage{
shipping_errors
}
\description{
blabla
}
\keyword{datasets}
