% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_times.R
\docType{data}
\name{sorting_times}
\alias{sorting_times}
\title{Sorting times}
\format{
A tibble with 60 rows and 1 variable:
\describe{
\item{sorting_time [dbl]}{Recorded time (in seconds) required by the machine for sorting individual parts.}
}
}
\usage{
sorting_times
}
\description{
The dataset contains throughput times of a machine recorded in order to compare them to the manufacturer's specifications.
}
\keyword{datasets}
