% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport_activity_by_occupation_individual.R
\docType{data}
\name{sport_activity_by_occupation_individual}
\alias{sport_activity_by_occupation_individual}
\title{Sport activity by occupation (individual-level)}
\format{
A tibble with 1,000 rows and 2 variables:
\describe{
\item{occupation}{Occupation category.}
\item{sport_activity}{Sport activity frequency category.}
}
}
\usage{
sport_activity_by_occupation_individual
}
\description{
The dataset contains one row per surveyed person with occupation and sport
activity frequency.
}
\keyword{datasets}
