% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport_occupation.R
\docType{data}
\name{sport_occupation}
\alias{sport_occupation}
\title{Sport activity by occupation (aggregated)}
\format{
A tibble with 5 rows and 4 variables:
\describe{
\item{occupation}{Occupation category.}
\item{never}{Count with no sport activity.}
\item{occasionally}{Count with occasional sport activity.}
\item{regularly}{Count with regular sport activity.}
}
}
\usage{
sport_occupation
}
\description{
The dataset contains counts of sport activity frequencies by occupation.
}
\keyword{datasets}
