% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_gender.R
\docType{data}
\name{tv_gender}
\alias{tv_gender}
\title{TV viewing time by gender (aggregated)}
\format{
A tibble with 3 rows and 4 variables:
\describe{
\item{tv_time_group}{Grouped weekly TV viewing time and total row.}
\item{men}{Count of men.}
\item{women}{Count of women.}
\item{total}{Row total.}
}
}
\usage{
tv_gender
}
\description{
The dataset contains grouped TV viewing counts by gender.
}
\keyword{datasets}
