% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_viewing_by_gender_individual.R
\docType{data}
\name{tv_viewing_by_gender_individual}
\alias{tv_viewing_by_gender_individual}
\title{TV viewing time by gender (individual-level)}
\format{
A tibble with 320 rows and 2 variables:
\describe{
\item{gender}{Gender of the surveyed person.}
\item{tv_time_group}{Grouped weekly TV viewing time.}
}
}
\usage{
tv_viewing_by_gender_individual
}
\description{
The dataset contains one row per person with gender and grouped weekly TV
viewing time.
}
\keyword{datasets}
