% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_agreement_satisfaction.R
\docType{data}
\name{union_agreement_satisfaction}
\alias{union_agreement_satisfaction}
\title{Satisfaction with a union agreement}
\format{
A tibble with 3 rows and 6 variables:
\describe{
\item{union}{Union identifier and total row.}
\item{very_satisfied}{Count of very satisfied respondents.}
\item{satisfied}{Count of satisfied respondents.}
\item{dissatisfied}{Count of dissatisfied respondents.}
\item{very_dissatisfied}{Count of very dissatisfied respondents.}
\item{total}{Row total.}
}
}
\usage{
union_agreement_satisfaction
}
\description{
The dataset contains a contingency table on member satisfaction with a new
agreement for two unions (plus totals).
}
\keyword{datasets}
