% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_satisfaction_ind.R
\docType{data}
\name{union_satisfaction_ind}
\alias{union_satisfaction_ind}
\title{Satisfaction with a union agreement (individual-level)}
\format{
A tibble with 1,000 rows and 2 variables:
\describe{
\item{union}{Union identifier.}
\item{satisfaction}{Satisfaction category for the agreement.}
}
}
\usage{
union_satisfaction_ind
}
\description{
The dataset contains the same information as
\link{union_satisfaction}, expanded to individual-level observations
with one row per surveyed person.
}
\keyword{datasets}
