% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{utility_vehicle_sales_weekly}
\alias{utility_vehicle_sales_weekly}
\title{Weekly utility vehicle sales}
\format{
A tibble with 6 rows and 2 variables:
\describe{
\item{vehicles_sold}{Number of vehicles sold per week.}
\item{probability}{Probability mass.}
}
}
\usage{
utility_vehicle_sales_weekly
}
\description{
Probability function for weekly units sold of utility vehicles.
}
\keyword{datasets}
