% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{vehicle_range_coefficients}
\alias{vehicle_range_coefficients}
\title{Vehicle range model coefficients}
\format{
A tibble with 4 rows and 2 variables:
\describe{
\item{regressor}{Regressor name.}
\item{coefficient}{Estimated coefficient.}
}
}
\usage{
vehicle_range_coefficients
}
\description{
Estimated coefficients from a multiple linear regression for vehicle range.
}
\keyword{datasets}
