% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{website_dwell_ind}
\alias{website_dwell_ind}
\title{Grouped website dwell times (individual-level)}
\format{
A tibble with 100 rows and 1 variable:
\describe{
\item{dwell_time_interval_min}{Interval of dwell time in minutes for the person.}
}
}
\usage{
website_dwell_ind
}
\description{
Expanded version of \link{website_dwell} with one row per sampled person.
}
\keyword{datasets}
