% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_dwell_time_grouped.R
\docType{data}
\name{website_dwell_time_grouped}
\alias{website_dwell_time_grouped}
\title{Grouped website dwell times}
\format{
A tibble with 5 rows and 2 variables:
\describe{
\item{dwell_time_interval_min}{Interval of dwell time in minutes.}
\item{persons}{Number of persons in the interval.}
}
}
\usage{
website_dwell_time_grouped
}
\description{
The dataset contains grouped dwell-time frequencies from a sample of 100
website users.
}
\keyword{datasets}
