% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{workshop_cars_per_hour}
\alias{workshop_cars_per_hour}
\title{Cars serviced per hour}
\format{
A tibble with 3 rows and 2 variables:
\describe{
\item{cars_per_hour}{Number of cars serviced per hour.}
\item{probability}{Probability mass.}
}
}
\usage{
workshop_cars_per_hour
}
\description{
Probability function for hourly number of cars serviced in a repair shop.
}
\keyword{datasets}
