% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_lecture_exercises.R
\docType{data}
\name{xy_quadratic}
\alias{xy_quadratic}
\title{Quadratic x-y observations}
\format{
A tibble with 7 rows and 3 variables:
\describe{
\item{i}{Observation index.}
\item{x}{Value of X.}
\item{y}{Value of Y.}
}
}
\usage{
xy_quadratic
}
\description{
Seven observations with a symmetric quadratic relationship between \code{x} and \code{y}.
}
\keyword{datasets}
