% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_data.R
\docType{data}
\name{hr_data}
\alias{hr_data}
\title{HR data}
\format{
A tibble with 1,200 rows and 6 variables:
\describe{
\item{position [fct]}{...}
\item{working_hours [dbl]}{...}
\item{salary [dbl]}{...}
\item{hourly_wage [dbl]}{...}
\item{department [fct]}{...}
\item{sick_days [dbl]}{...}
}
}
\usage{
hr_data
}
\description{
Edit description here.
}
\keyword{datasets}
