% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_expenses.R
\docType{data}
\name{marketing_expenses}
\alias{marketing_expenses}
\title{Data of Marketing expenses}
\format{
A tibble with 235 rows and 14 variables:
\describe{
\item{marketing_expenses [dbl]}{...}
\item{customers_reached [dbl]}{...}
\item{negative_reactions [dbl]}{...}
\item{price [dbl]}{...}
\item{price_segment [fct]}{...}
\item{number_of_sizes [dbl]}{...}
\item{target_customer [fct]}{...}
\item{rating_testers [dbl]}{...}
\item{rating_customers [dbl]}{...}
\item{color_most_sold [fct]}{...}
\item{return_rate [dbl]}{...}
\item{sales_volume [dbl]}{...}
\item{rank_rating_customers [dbl]}{...}
\item{rank_price_segment [dbl]}{...}
}
}
\usage{
marketing_expenses
}
\description{
Marketing expenses for shoes.
}
\keyword{datasets}
