% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSBStatsData-package.R
\docType{package}
\name{MSBStatsData-package}
\alias{MSBStatsData}
\alias{MSBStatsData-package}
\title{MSBStatsData: Data Sets for Courses at the M端nster School of Business}
\description{
Provides sample data sets that are used in statistics and data science courses at the M端nster School of Business. The datasets refer to different business topics but also other domains, e.g. sports, traffic, etc.
}
\author{
\strong{Maintainer}: Michael B端cker \email{michael.buecker@fh-muenster.de} (\href{https://orcid.org/0000-0003-0045-8460}{ORCID})

Authors:
\itemize{
  \item Niels Schl端sener \email{niels.schluesener@fh-muenster.de}
}

}
\keyword{internal}
