% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipping_errors.R
\docType{data}
\name{shipping_errors}
\alias{shipping_errors}
\title{Shipping error data}
\format{
A tibble with 576 rows and 2 variables:
\describe{
\item{error [fct]}{Indicator whether or not an error appeared during shipment.}
\item{shipping [fct]}{Shipment method requested by the customer.}
}
}
\usage{
shipping_errors
}
\description{
Dataset recording errors in the shipment and the type of delivery requested by the customer in the process.
}
\keyword{datasets}
