% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_times.R
\docType{data}
\name{sorting_times}
\alias{sorting_times}
\title{Sorting times}
\format{
A tibble with 10 rows and 2 variables:
\describe{
\item{sorting_time [dbl]}{...}
}
}
\usage{
sorting_times
}
\description{
blabla
}
\keyword{datasets}
